Steal the main thread on OSX
Should fix issues with events not being handled properly, and
possibly other things. Taken from Erlang's wx, which is proven
to work. The function used is unfortunately undocumented.

Not 100% sure everything is good with just that, it looks good
enough on my VM, but I still have issues probably due to trying
this on a VM.