Add sdl_gl:create_context/1 and sdl_gl:swap_window/1
These functions are the bare minimum needed to experiment
with OpenGL.

My experiments so far rule out reusing the gl and glu modules
found in the wx application. It seems that any OpenGL call
from a different thread than the window's thread won't work,
at least on Linux.

I'm afraid that implementing OpenGL in the future will require
duplicating the work done in the wx application. OpenGL support
is not a priority however, so it's not a big deal just yet.